// services/attendance/src/lib/database/models/attendance.model.ts
export enum AttendanceStatus {
  PRESENT = 'PRESENT',
  ABSENT = 'ABSENT',
  LATE = 'LATE',
  EARLY_LEAVE = 'EARLY_LEAVE',
  HOLIDAY = 'HOLIDAY',
  SICK_LEAVE = 'SICK_LEAVE',
  VACATION = 'VACATION',
  REMOTE = 'REMOTE',
  BUSINESS_TRIP = 'BUSINESS_TRIP',
}